fstests: UDEV_SETTLE_PROG before dmsetup create
We've been seeing transient errors with any test that uses a dm device
for the entirety of the time that we've been running nightly xfstests
runs.  This turns out to be because sometimes we get EBUSY while trying
to create our new dm device.  Generally this is because the test comes
right after another test that messes with the dm device, and thus we
still have udev messing around with the device when DM tries to O_EXCL
the block device.

Add a UDEV_SETTLE_PROG before creating the device to make sure we can
create our new dm device without getting this transient error.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>